Refactor ServiceOrchestrator
Split `ServiceOrchestrator.ts` into smaller, more manageable files to improve code organization and maintainability.
